chore(cairo): merged this branch with Patrick's changes and refactored data copying

+++ core/src/main/c/windows/files.c
@@ -357,7 +357,7 @@ inline jlong _io_questdb_std_Files_mremap0
     jlong newAddress = Java_io_questdb_std_Files_mmap0((JNIEnv *) NULL, (jclass) NULL, fd, newLen, offset, flags);
     // Note that unmapping will not flush dirty pages because the mapping to address is shared with newAddress
     Java_io_questdb_std_Files_munmap0((JNIEnv *) NULL, (jclass) NULL, address, previousLen);
-    return (jlong) newAddress;
+    return newAddress;
 }
     
 JNIEXPORT jlong JNICALL JavaCritical_io_questdb_std_Files_mremap0
